backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Data validation
pydantic==2.7.1

# Database
sqlalchemy==2.0.30

# Audio / DSP
torch>=2.2.0
torchaudio>=2.2.0
transformers>=4.40.0
faster-whisper>=1.0.0
librosa>=0.10.0
soundfile>=0.12.1
scipy>=1.13.0
numpy>=1.26.0
pydub>=0.25.1

# NLP / ML
scikit-learn>=1.4.0
nltk>=3.8.1         # CMUDict for syllable counting
python-Levenshtein>=0.25.0

# Utilities
python-dotenv==1.0.1
aiofiles==23.2.1
